requirements.txt
numpy
pandas
nltk
scikit-learn
xgboost
fastapi
uvicorn
pydantic
joblib
requests
deepl
python-dotenv